requirements.txt
# Essential Python packages for Stock Market Prediction ML Project

# Data Collection and Processing
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.15.0

# Technical Analysis (alternative to talib)
pandas-ta>=0.3.14b0

# Sentiment Analysis
vaderSentiment>=3.3.2
textblob>=0.17.1

# Visualization and Dashboard
plotly>=5.17.0
streamlit>=1.28.0

# Data Science Tools
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Natural Language Processing
textblob==0.17.1
vaderSentiment==3.3.2
nltk==3.8.1
transformers==4.36.2

# Visualization and Dashboard
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
streamlit==1.29.0
dash==2.16.1

# Data Storage and Configuration
sqlite3
pyyaml==6.0.1
python-dotenv==1.0.0

# API and Web Scraping
newsapi-python==0.2.6
tweepy==4.14.0
alpha-vantage==2.3.1

# Utilities
tqdm==4.66.1
joblib==1.3.2
python-dateutil==2.8.2

# Testing and Development
pytest==7.4.3
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1

# Performance and Optimization
numba==0.58.1
cython==3.0.6
